Rewrote all-the-things to use prototypes and refactored existing code
This commit is pretty huge as it replaces every line in quay.js. The reason for this is because
originally Quay was just a hack to get something I was working on done quickly. I've been meaning
to rewrite it to use prototypes and clean up the code for a long time. This commit doesn't change any
functionality at all but refactors all-the-codes.

Also to note:
	* Array.prototype.filter isn't supported in <IE9 so I've added a fallback to keep backwards
compatibility.